# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Base Next.js API",
    "description": "API de autenticação e gerenciamento de usuários",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Endpoints de autenticação"
    },
    {
      "name": "Usuarios",
      "description": "Endpoints de gerenciamento de usuários"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Autenticar usuário",
        "description": "Autentica o usuário e define o cookie HTTP-only com o token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "senha"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email do usuário"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "description": "Senha do usuário"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": {
                      "$ref": "#/components/schemas/UsuarioResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Fazer logout",
        "description": "Faz logout do usuário, limpando o cookie de token",
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Logout realizado com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "post": {
        "tags": ["Auth"],
        "summary": "Obter usuário atual",
        "description": "Retorna os dados do usuário atual baseado no token do cookie",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": {
                      "$ref": "#/components/schemas/UsuarioResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar usuários",
        "description": "Lista todos os usuários (exclui soft-deleted)",
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuarios": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UsuarioListItem"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Criar usuário",
        "description": "Cria um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "email", "senha"],
                "properties": {
                  "nome": {
                    "type": "string",
                    "description": "Nome do usuário"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email do usuário"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "description": "Senha do usuário"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": {
                      "$ref": "#/components/schemas/UsuarioResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Buscar usuário por ID",
        "description": "Busca um usuário específico pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": {
                      "$ref": "#/components/schemas/UsuarioResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Usuarios"],
        "summary": "Atualizar usuário",
        "description": "Atualiza um usuário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "description": "Nome do usuário"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email do usuário"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "description": "Nova senha do usuário"
                  },
                  "senhaAtual": {
                    "type": "string",
                    "format": "password",
                    "description": "Senha atual (obrigatória para alterar a senha)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": {
                      "$ref": "#/components/schemas/UsuarioResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Senha atual incorreta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Excluir usuário",
        "description": "Exclui um usuário (soft delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Usuário excluído com sucesso"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
      "UsuarioResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do usuário"
          },
          "nome": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "description": "Data de atualização"
          }
        }
      },
      "UsuarioListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do usuário"
          },
          "nome": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string",
            "description": "Mensagem de erro"
          }
        }
      }
    }
  }
}